Modify order of commands in generate_categorical_fim_flows and place input paths from env file into a function. Rename generate_categorical_fim_mapping. Add wrapper script to run both categorical fim scripts.
